Might fail, not entirely sure.
Former-commit-id: 34736945dd85961d0795bc242586c8b6ff49962e [formerly ba04b0649f5b86bb14177dea776243236cc87cce]
Former-commit-id: 27001e24e57684a15d4bac7099955536d5382823

+++ src/librapid/math/advanced.hpp
@@ -63,7 +63,7 @@ namespace librapid {
 					x += std::complex<double>(pVal) /
 						 (z + std::complex<double>(i) + std::complex<double>(1));
 				}
-				auto t = z + std::complex<double>(elemsP) - std::complex<double>(0.5);
+				auto t = z + std::complex<double>((double)elemsP) - std::complex<double>(0.5);
 				y	   = sqrt(2 * PI) * pow(t, z + 0.5) * exp(-t) * x;
 			}
